Combine EmptyCharset test with IllegalCharset test
